add test to fuselage

+++ src/test/java/inf112/skeleton/model/SpaceGameModelTest.java
@@ -71,8 +71,9 @@ class SpaceGameModelTest {
         assertNotNull(laser);
 
         gameModel.moveLaser();
-        assertEquals(initialPlayerY + gameModel.getSpaceShips()[0].getShipStructure().getHeight() / 2 + 1,
-                laser.getY());
+        assertEquals(initialPlayerY + (float) gameModel.getSpaceShips()[0].getShipStructure()
+                .getHeight() / 2 + 1,
+            laser.getY());
 
         // A laser shoots from a ship, and is deleted when offscreen, so
         // get the x coordinate of the laser for this loop.
